Updates to glint360k loader data
Calculating an accurate time when starting the multi-process in python is simply not in the cards, I tried using a `chunk_size` of 1 and use more parallelization but that will simply be too slow. In order to keep it fast this commit only outputs an increasing percent number for the first worker thread.
For better understanding it will also show the time passed for every log statement.